plumbing params through configs

+++ test/test_sinkd.py
@@ -7,11 +7,13 @@ from pathlib import Path
 import multiprocessing as mp
 
 
-TLD = None
-CLIENT_PATH = None
-SERVER_PATH = None
-SYSTEM_CFG = None
-CLIENT_CFG = None
+TLD = Path(__file__).parents[1]
+SYSTEM_CFG = TLD.joinpath("test", "etc_sinkd.conf")
+USER_CFG = TLD.joinpath("test", "sinkd.conf")
+
+ROOT_PATH = Path(TLD, "test", "sinkd_dmz")
+CLIENT_PATH = Path(ROOT_PATH, "client")
+SERVER_PATH = Path(ROOT_PATH, "server")
 
 
 def run(cmd, **kwargs) -> subprocess.CompletedProcess:
@@ -22,34 +24,19 @@ def run(cmd, **kwargs) -> subprocess.CompletedProcess:
 
 
 def setup_env():
-    global TLD, CLIENT_PATH, SERVER_PATH, SYSTEM_CFG, USER_CFG
-
-    curr_path = Path(__file__)
-    test_dir = curr_path.parent
-    TLD = curr_path.parents[1]
-
-    SYSTEM_CFG = test_dir.joinpath("etc_sinkd.conf")
-    USER_CFG = test_dir.joinpath("sinkd.conf")
-
-    print("system cfg path: ", SYSTEM_CFG)
-    print("user cfg path: ", USER_CFG)
-
-    ROOT_PATH = Path(TLD, "test", "sinkd_dmz")
-    CLIENT_PATH = Path(ROOT_PATH, "client")
-    SERVER_PATH = Path(ROOT_PATH, "server")
-
     ROOT_PATH.mkdir(exist_ok=True, parents=True)
     CLIENT_PATH.mkdir(exist_ok=True)
     SERVER_PATH.mkdir(exist_ok=True)
 
 
 def remove_subfiles(directory: Path):
-    for f in directory.glob("*"):
-        try:
-            shutil.rmtree(f)
-            print("removed ", f)
-        except FileNotFoundError as e:
-            print(f"File not found: {e}")
+    if directory:
+        for f in directory.glob("*"):
+            try:
+                shutil.rmtree(f)
+                print("removed ", f)
+            except FileNotFoundError as e:
+                print(f"File not found: {e}")
 
 
 def create_files(folder: Path, num_of_files: int, delay: float = 0.01):
@@ -66,7 +53,7 @@ def create_files(folder: Path, num_of_files: int, delay: float = 0.01):
 def spawn_sinkd():
     client = run(f"./target/debug/sinkd --debug -s {SYSTEM_CFG} -u {USER_CFG} start --client")
     if client.returncode != 0:
-        print("uh oh", client.stderr, client.stdout)
+        print("test_sinkd>> ", client.stderr, client.stdout)
         exit(-1)
     print("sucessfully spawned sinkd")
